Hierarchical clustering of differentially expressed genes relative to a baseline sample.
In order to be included, genes must be expressed highly (>100 FPKM) in
at least one sample and must be expressed 2-fold above or below the
reference sample (combined immature neurons). Fold change must meet a
q-value of 0.1 to be considered significant. This q-value is calculated
from pairwise (sample1 vs sample2) ANOVAs on every gene. The q-value is
an FDR-corrected p-value.